BPEL updated with Helper
Please note: I have not tested it yet, cause there is an issue. I have A
flow where two for-each loops going (one for flights another for
hotels), then inside there are if condition IF (dateDeparture[$i] is
less than closestDate (which is set to currentDate)) then
closestDateFlight = dateDeparture[$i];. the same is done with hotels.
after this flow there is assign to currentDate. and if condition if
flightClosestDate is less than hotelClosestDate there are assigns Done -
closestDate = flightClosestDate or ELSE condition closestDate =
hotelClosestDate. After this we have talked with johannes to add one
more IF which is checking IF (closestDate is less than currentDate) then
the currentDate is equal to closestDate; This is done because if the
date is in the future = itinerary is waiting until that date. But i have
not tested it because NOW i am not able to run tests normally - it keeps
going. I am commiting now, but please someone if have time - have a look
at this. MAYBE you will notice my mistake, and will help me to fix
it.... Cause I really do not have time to do that at the moment..
Hopefully I will be able to join you tomorrow, but i will let you know
later on..